improvements to ognode/laser2og.
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@3002 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/components/pathplanner/genericdriver.cpp
@@ -53,7 +53,7 @@ namespace {
 
                 if ( orcapathplan::isTraversable( ogMap, cell.x(), cell.y(), traversabilityThreshhold ) )
                 {
-                    float worldX, worldY;
+                    double worldX, worldY;
                     ogMap.getWorldCoords( cell.x(), cell.y(), worldX, worldY );
                     wp.target.p.x = worldX;
                     wp.target.p.y = worldY;
